% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_psychrometric_computations.R
\name{calc_psychrometrics_from_rel_hum}
\alias{calc_psychrometrics_from_rel_hum}
\title{Calculate psychrometric values from relative humidity.}
\usage{
calc_psychrometrics_from_rel_hum(t_dry_bulb, rel_hum, pressure)
}
\arguments{
\item{t_dry_bulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{rel_hum}{Relative humidity in range [0, 1]}

\item{pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Vector with named components for each psychrometric value computed:
  \describe{
    \item{hum_ratio}{Humidity ratio in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]}
    \item{t_wet_bulb}{Wet-bulb temperature in °F [IP] or °C [SI]}
    \item{t_dew_point}{Dew-point temperature in °F [IP] or °C [SI]}
    \item{vap_pres}{Partial pressure of water vapor in moist air in Psi [IP] or Pa [SI]}
    \item{moist_air_enthalpy}{Moist air enthalpy in Btu lb⁻¹ [IP] or J kg⁻¹ [SI]}
    \item{moist_air_volume}{Specific volume of moist air in ft³ lb⁻¹ [IP] or in m³ kg⁻¹ [SI]}
    \item{degree_of_saturation}{Degree of saturation [unitless]}
  }
}
\description{
Utility function to calculate humidity ratio, wet-bulb temperature, dew-point temperature,
 vapour pressure, moist air enthalpy, moist air volume, and degree of saturation of air given
 dry-bulb temperature, relative humidity and pressure.
}
