% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk.R
\name{CalcPsychrometricsFromTWetBulb}
\alias{CalcPsychrometricsFromTWetBulb}
\title{Calculate psychrometric values from wet-bulb temperature.}
\usage{
CalcPsychrometricsFromTWetBulb(TDryBulb, TWetBulb, Pressure)
}
\arguments{
\item{TDryBulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{TWetBulb}{Wet-bulb temperature in °F [IP] or °C [SI]}

\item{Pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Vector with named components for each psychrometric value computed:
\describe{
  \item{HumRatio}{Humidity ratio in lb_H2O lb_Air-1 [IP] or kg_H2O kg_Air-1 [SI]}
  \item{TDewPoint}{Dew-point temperature in °F [IP] or °C [SI]}
  \item{RelHum}{Relative humidity in range [0, 1]}
  \item{VapPres}{Partial pressure of water vapor in moist air in Psi [IP] or Pa [SI]}
  \item{MoistAirEnthalpy}{Moist air enthalpy in Btu lb-1 [IP] or J kg-1 [SI]}
  \item{MoistAirVolume}{Specific volume of moist air in ft3 lb-1 [IP] or in m3 kg-1 [SI]}
  \item{DegreeOfSaturation}{Degree of saturation [unitless]}
}
}
\description{
Utility function to calculate humidity ratio, dew-point temperature, relative humidity,
vapour pressure, moist air enthalpy, moist air volume, and degree of saturation of air given
dry-bulb temperature, wet-bulb temperature, and pressure.
}
