% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dew_point.R
\name{dLnPws_}
\alias{dLnPws_}
\title{Helper function returning the derivative of the natural log of the saturation vapor pressure
as a function of dry-bulb temperature.}
\usage{
dLnPws_(TDryBulb)
}
\arguments{
\item{TDryBulb}{A numeric vector of dry-bulb temperature in degreeF [IP] or degreeC [SI]}
}
\value{
A numeric vector of derivative of natural log of vapor pressure of saturated air in Psi [IP] or Pa [SI]
}
\description{
Helper function returning the derivative of the natural log of the saturation vapor pressure
as a function of dry-bulb temperature.
}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1  eqn 5 & 6
}
\keyword{internal}
