% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moist_air_calculations.R
\name{get_degree_of_saturation}
\alias{get_degree_of_saturation}
\title{Return the degree of saturation (i.e humidity ratio of the air / humidity ratio of the air at saturation
at the same temperature and pressure) given dry-bulb temperature, humidity ratio, and atmospheric pressure.}
\usage{
get_degree_of_saturation(t_dry_bulb, hum_ratio, pressure)
}
\arguments{
\item{t_dry_bulb}{numeric Dry-bulb temperature in °F [IP] or °C [SI]}

\item{hum_ratio}{numeric Humidity ratio in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]}

\item{pressure}{numeric Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
numeric Degree of saturation in arbitrary unit

Reference:
  ASHRAE Handbook - Fundamentals (2009) ch. 1 eqn 12

Notes:
  This definition is absent from the 2017 Handbook. Using 2009 version instead.
}
\description{
Return the degree of saturation (i.e humidity ratio of the air / humidity ratio of the air at saturation
at the same temperature and pressure) given dry-bulb temperature, humidity ratio, and atmospheric pressure.
}
