% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_air_calculations.R
\name{get_dry_air_density}
\alias{get_dry_air_density}
\title{Return dry-air density given dry-bulb temperature and pressure.}
\usage{
get_dry_air_density(t_dry_bulb, pressure)
}
\arguments{
\item{t_dry_bulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Dry air density in lb ft⁻³ [IP] or kg m⁻³ [SI]
}
\description{
Return dry-air density given dry-bulb temperature and pressure.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1
  \itemize{
    \item Eqn 14 for the perfect gas relationship for dry air.
    \item Eqn 1 for the universal gas constant.
  }
}

