% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_air_calculations.R
\name{get_hum_ratio_from_enthalpy_and_t_dry_bulb}
\alias{get_hum_ratio_from_enthalpy_and_t_dry_bulb}
\title{Return humidity ratio from enthalpy and dry-bulb temperature.}
\usage{
get_hum_ratio_from_enthalpy_and_t_dry_bulb(moist_air_enthalpy, t_dry_bulb)
}
\arguments{
\item{moist_air_enthalpy}{Moist air enthalpy in Btu lb⁻¹ [IP] or J kg⁻¹}

\item{t_dry_bulb}{Dry-bulb temperature in °F [IP] or °C [SI]}
}
\value{
Humidity ratio in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]
}
\description{
Return humidity ratio from enthalpy and dry-bulb temperature.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 30.
}

