% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/wet_bulb_dew_point_or_rel_hum_to_humidity_ratio_conversions.R
\name{get_hum_ratio_from_rel_hum}
\alias{get_hum_ratio_from_rel_hum}
\title{Return humidity ratio given dry-bulb temperature, relative humidity, and pressure.}
\usage{
get_hum_ratio_from_rel_hum(t_dry_bulb, rel_hum, pressure)
}
\arguments{
\item{t_dry_bulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{rel_hum}{Relative humidity in range [0, 1]}

\item{pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Humidity ratio in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]
}
\description{
Return humidity ratio given dry-bulb temperature, relative humidity, and pressure.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1
}

