% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/humidity_ratio_and_specific_humidity_conversions.R
\name{get_hum_ratio_from_specific_hum}
\alias{get_hum_ratio_from_specific_hum}
\title{Return the humidity ratio (aka mixing ratio) from specific humidity.}
\usage{
get_hum_ratio_from_specific_hum(specific_hum)
}
\arguments{
\item{specific_hum}{Specific humidity in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]}
}
\value{
Humidity ratio in lb_H₂O lb_Dry_Air⁻¹ [IP] or kg_H₂O kg_Dry_Air⁻¹ [SI]
}
\description{
Return the humidity ratio (aka mixing ratio) from specific humidity.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 9b (solved for humidity ratio)
}

