% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/wet_bulb_dew_point_or_rel_hum_to_humidity_ratio_conversions.R
\name{get_hum_ratio_from_t_dew_point}
\alias{get_hum_ratio_from_t_dew_point}
\title{Return humidity ratio given dew-point temperature and pressure.}
\usage{
get_hum_ratio_from_t_dew_point(t_dew_point, pressure)
}
\arguments{
\item{t_dew_point}{Dew-point temperature in °F [IP] or °C [SI]}

\item{pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Humidity ratio in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]
}
\description{
Return humidity ratio given dew-point temperature and pressure.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 13
}

