% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/wet_bulb_dew_point_or_rel_hum_to_humidity_ratio_conversions.R
\name{get_hum_ratio_from_t_wet_bulb}
\alias{get_hum_ratio_from_t_wet_bulb}
\title{Return humidity ratio given dry-bulb temperature, wet-bulb temperature, and pressure.}
\usage{
get_hum_ratio_from_t_wet_bulb(t_dry_bulb, t_wet_bulb, pressure)
}
\arguments{
\item{t_dry_bulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{t_wet_bulb}{Wet-bulb temperature in °F [IP] or °C [SI]}

\item{pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Humidity ratio in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]
}
\description{
Return humidity ratio given dry-bulb temperature, wet-bulb temperature, and pressure.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 33 and 35
}

