% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity_ratio_and_vapor_pressure_conversions.R
\name{get_hum_ratio_from_vap_pres}
\alias{get_hum_ratio_from_vap_pres}
\title{Return humidity ratio given water vapor pressure and atmospheric pressure.}
\usage{
get_hum_ratio_from_vap_pres(vap_pres, pressure)
}
\arguments{
\item{vap_pres}{Partial pressure of water vapor in moist air in Psi [IP] or Pa [SI]}

\item{pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Humidity ratio in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]
}
\description{
Return humidity ratio given water vapor pressure and atmospheric pressure.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 20
}

