% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moist_air_calculations.R
\name{get_moist_air_enthalpy}
\alias{get_moist_air_enthalpy}
\title{Return moist air enthalpy given dry-bulb temperature and humidity ratio.}
\usage{
get_moist_air_enthalpy(t_dry_bulb, hum_ratio)
}
\arguments{
\item{t_dry_bulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{hum_ratio}{Humidity ratio in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]}
}
\value{
Moist air enthalpy in Btu lb⁻¹ [IP] or J kg⁻¹
}
\description{
Return moist air enthalpy given dry-bulb temperature and humidity ratio.
}
\section{Reference}{

 ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 30
}

