% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moist_air_calculations.R
\name{get_moist_air_volume}
\alias{get_moist_air_volume}
\title{Return moist air specific volume given dry-bulb temperature, humidity ratio, and pressure.}
\usage{
get_moist_air_volume(t_dry_bulb, hum_ratio, pressure)
}
\arguments{
\item{t_dry_bulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{hum_ratio}{Humidity ratio in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]}

\item{pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Specific volume of moist air in ft³ lb⁻¹ of dry air [IP] or in m³ kg⁻¹ of dry air [SI]
}
\description{
Return moist air specific volume given dry-bulb temperature, humidity ratio, and pressure.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 26
}

