% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dew_point_wet_bulb_relative_humidity_conversions.R
\name{get_rel_hum_from_t_dew_point}
\alias{get_rel_hum_from_t_dew_point}
\title{Return relative humidity given dry-bulb temperature and dew-point temperature.}
\usage{
get_rel_hum_from_t_dew_point(t_dry_bulb, t_dew_point)
}
\arguments{
\item{t_dry_bulb}{Dry-bulb temperature in 째F [IP] or 째C [SI]}

\item{t_dew_point}{Dew-point temperature in 째F [IP] or 째C [SI]}
}
\value{
Relative humidity in range [0, 1]
}
\description{
Return relative humidity given dry-bulb temperature and dew-point temperature.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 22
}

