% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dew_point_wet_bulb_relative_humidity_conversions.R
\name{get_rel_hum_from_t_wet_bulb}
\alias{get_rel_hum_from_t_wet_bulb}
\title{Return relative humidity given dry-bulb temperature, wet bulb temperature and pressure.}
\usage{
get_rel_hum_from_t_wet_bulb(t_dry_bulb, t_wet_bulb, pressure)
}
\arguments{
\item{t_dry_bulb}{Dry-bulb temperature in 째F [IP] or 째C [SI]}

\item{t_wet_bulb}{Wet-bulb temperature in 째F [IP] or 째C [SI]}

\item{pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Relative humidity in range [0, 1]
}
\description{
Return relative humidity given dry-bulb temperature, wet bulb temperature and pressure.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1
}

