% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturated_air_calculations.R
\name{get_sat_vap_pres}
\alias{get_sat_vap_pres}
\title{Return saturation vapor pressure given dry-bulb temperature.}
\usage{
get_sat_vap_pres(t_dry_bulb)
}
\arguments{
\item{t_dry_bulb}{Dry-bulb temperature in °F [IP] or °C [SI]}
}
\value{
Vapor pressure of saturated air in Psi [IP] or Pa [SI]
}
\description{
Return saturation vapor pressure given dry-bulb temperature.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1  eqn 5 & 6
}

