% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_atmosphere_calculations.R
\name{get_sea_level_pressure}
\alias{get_sea_level_pressure}
\title{Return sea level pressure given dry-bulb temperature, altitude above sea level and pressure.}
\usage{
get_sea_level_pressure(station_pressure, altitude, t_dry_bulb)
}
\arguments{
\item{station_pressure}{Observed station pressure in Psi [IP] or Pa [SI]}

\item{altitude}{Altitude in ft [IP] or m [SI]}

\item{t_dry_bulb}{Dry-bulb temperature in °F [IP] or °C [SI]}
}
\value{
Sea level barometric pressure in Psi [IP] or Pa [SI]
}
\description{
Return sea level pressure given dry-bulb temperature, altitude above sea level and pressure.
}
\section{Reference}{

  Hess SL, Introduction to theoretical meteorology, Holt Rinehart and Winston, NY 1959,
  ch. 6.5; Stull RB, Meteorology for scientists and engineers, 2nd edition,
  Brooks/Cole 2000, ch. 1.
}

\section{Notes}{

 The standard procedure for the US is to use for TDryBulb the average
 of the current station temperature and the station temperature from 12 hours ago.
}

