% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/humidity_ratio_and_specific_humidity_conversions.R
\name{get_specific_hum_from_hum_ratio}
\alias{get_specific_hum_from_hum_ratio}
\title{Return the specific humidity from humidity ratio (aka mixing ratio).}
\usage{
get_specific_hum_from_hum_ratio(hum_ratio)
}
\arguments{
\item{hum_ratio}{Humidity ratio in lb_H₂O lb_Dry_Air⁻¹ [IP] or kg_H₂O kg_Dry_Air⁻¹ [SI]}
}
\value{
Specific humidity in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]
}
\description{
Return the specific humidity from humidity ratio (aka mixing ratio).
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 9b
}

