% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_atmosphere_calculations.R
\name{get_standard_atm_pressure}
\alias{get_standard_atm_pressure}
\title{Return standard atmosphere barometric pressure, given the elevation (altitude).}
\usage{
get_standard_atm_pressure(altitude)
}
\arguments{
\item{altitude}{numeric Altitude in ft [IP] or m [SI]}
}
\value{
numeric Standard atmosphere barometric pressure in Psi [IP] or Pa [SI]

Reference:
  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 3
}
\description{
Return standard atmosphere barometric pressure, given the elevation (altitude).
}
