% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_atmosphere_calculations.R
\name{get_standard_atm_temperature}
\alias{get_standard_atm_temperature}
\title{Return standard atmosphere temperature, given the elevation (altitude).}
\usage{
get_standard_atm_temperature(altitude)
}
\arguments{
\item{altitude}{Altitude in ft [IP] or m [SI]}
}
\value{
Standard atmosphere dry-bulb temperature in °F [IP] or °C [SI]
}
\description{
Return standard atmosphere temperature, given the elevation (altitude).
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 4
}

