% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_atmosphere_calculations.R
\name{get_station_pressure}
\alias{get_station_pressure}
\title{Return station pressure from sea level pressure.}
\usage{
get_station_pressure(sea_level_pressure, altitude, t_dry_bulb)
}
\arguments{
\item{sea_level_pressure}{Sea level barometric pressure in Psi [IP] or Pa [SI]}

\item{altitude}{Altitude in ft [IP] or m [SI]}

\item{t_dry_bulb}{Dry-bulb temperature in °F [IP] or °C [SI]}
}
\value{
Station pressure in Psi [IP] or Pa [SI]
}
\description{
Return station pressure from sea level pressure.
}
\section{Reference}{

  See \code{\link{get_sea_level_pressure}}.
}

\section{Notes}{

  This function is just the inverse of \code{\link{get_sea_level_pressure}}.
}

