% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dew_point_wet_bulb_relative_humidity_conversions.R
\name{get_t_dew_point_from_rel_hum}
\alias{get_t_dew_point_from_rel_hum}
\title{Return dew-point temperature given dry-bulb temperature and relative humidity.}
\usage{
get_t_dew_point_from_rel_hum(t_dry_bulb, rel_hum)
}
\arguments{
\item{t_dry_bulb}{Dry-bulb temperature in 째F [IP] or 째C [SI]}

\item{rel_hum}{Relative humidity in range [0, 1]}
}
\value{
Dew-point temperature in 째F [IP] or 째C [SI]
}
\description{
Return dew-point temperature given dry-bulb temperature and relative humidity.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1
}

