% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dew_point_wet_bulb_relative_humidity_conversions.R
\name{get_t_dew_point_from_t_wet_bulb}
\alias{get_t_dew_point_from_t_wet_bulb}
\title{Return dew-point temperature given dry-bulb temperature, wet-bulb temperature, and pressure.}
\usage{
get_t_dew_point_from_t_wet_bulb(t_dry_bulb, t_wet_bulb, pressure)
}
\arguments{
\item{t_dry_bulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{t_wet_bulb}{Wet-bulb temperature in °F [IP] or °C [SI]}

\item{pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Dew-point temperature in °F [IP] or °C [SI]
}
\description{
Return dew-point temperature given dry-bulb temperature, wet-bulb temperature, and pressure.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1
}

