% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dew_point_or_relative_humidity_and_vapor_pressure_conversions.R
\name{get_t_dew_point_from_vap_pres}
\alias{get_t_dew_point_from_vap_pres}
\title{Return dew-point temperature given dry-bulb temperature and vapor pressure.}
\usage{
get_t_dew_point_from_vap_pres(t_dry_bulb, vap_pres)
}
\arguments{
\item{t_dry_bulb}{Dry-bulb temperature in 째F [IP] or 째C [SI]}

\item{vap_pres}{Partial pressure of water vapor in moist air in Psi [IP] or Pa [SI]}
}
\value{
Dew-point temperature in 째F [IP] or 째C [SI]
}
\description{
Return dew-point temperature given dry-bulb temperature and vapor pressure.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn. 5 and 6
}

\section{Notes}{

\itemize{
 \item The dew point temperature is solved by inverting the equation giving water vapor pressure
 at saturation from temperature rather than using the regressions provided
 by ASHRAE (eqn. 37 and 38), which are much less accurate and have a
 narrower range of validity.
 \item The Newton-Raphson (NR) method is used on the logarithm of water vapour
 pressure as a function of temperature, which is a very smooth function.
 \item Convergence is usually achieved in 3 to 5 iterations.
 \item t_dry_bulb is not really needed here, just used for convenience.
}
}

