% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_air_calculations.R
\name{get_t_dry_bulb_from_enthalpy_and_hum_ratio}
\alias{get_t_dry_bulb_from_enthalpy_and_hum_ratio}
\title{Return dry bulb temperature from enthalpy and humidity ratio.}
\usage{
get_t_dry_bulb_from_enthalpy_and_hum_ratio(moist_air_enthalpy, hum_ratio)
}
\arguments{
\item{moist_air_enthalpy}{numeric Moist air enthalpy in Btu lb⁻¹ [IP] or J kg⁻¹}

\item{hum_ratio}{numeric Humidity ratio in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]}
}
\value{
numeric Dry-bulb temperature in °F [IP] or °C [SI]

Reference:
  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 30

Notes:
  Based on the `GetMoistAirEnthalpy` function, rearranged for temperature.
}
\description{
Return dry bulb temperature from enthalpy and humidity ratio.
}
