% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_conversions.R
\name{get_t_rankine_from_t_fahrenheit}
\alias{get_t_rankine_from_t_fahrenheit}
\title{Utility function to convert temperature to degree Rankine (°R)
given temperature in degree Fahrenheit (°F).}
\usage{
get_t_rankine_from_t_fahrenheit(t_fahrenheit)
}
\arguments{
\item{t_fahrenheit}{Temperature in degree Fahrenheit (°F)}
}
\value{
Temperature in degree Rankine (°R)
}
\description{
Utility function to convert temperature to degree Rankine (°R)
given temperature in degree Fahrenheit (°F).
}
\section{Notes}{

  Exact conversion.
}

