% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/wet_bulb_dew_point_or_rel_hum_to_humidity_ratio_conversions.R
\name{get_t_wet_bulb_from_hum_ratio}
\alias{get_t_wet_bulb_from_hum_ratio}
\title{Return wet-bulb temperature given dry-bulb temperature, humidity ratio, and pressure.}
\usage{
get_t_wet_bulb_from_hum_ratio(t_dry_bulb, hum_ratio, pressure)
}
\arguments{
\item{t_dry_bulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{hum_ratio}{Humidity ratio in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]}

\item{pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Wet-bulb temperature in °F [IP] or °C [SI]
}
\description{
Return wet-bulb temperature given dry-bulb temperature, humidity ratio, and pressure.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 33 and 35 solved for Tstar
}

