% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_unit_system}
\alias{get_unit_system}
\title{Return system of units in use.}
\usage{
get_unit_system()
}
\value{
string indicating system of units in use ("SI" or "IP")
}
\description{
Return system of units in use.
}
