% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity_ratio_and_vapor_pressure_conversions.R
\name{get_vap_pres_from_hum_ratio}
\alias{get_vap_pres_from_hum_ratio}
\title{Return vapor pressure given humidity ratio and pressure.}
\usage{
get_vap_pres_from_hum_ratio(hum_ratio, pressure)
}
\arguments{
\item{hum_ratio}{Humidity ratio in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]}

\item{pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Partial pressure of water vapor in moist air in Psi [IP] or Pa [SI]
}
\description{
Return vapor pressure given humidity ratio and pressure.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 20 solved for pw
}

