% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dew_point_or_relative_humidity_and_vapor_pressure_conversions.R
\name{get_vap_pres_from_rel_hum}
\alias{get_vap_pres_from_rel_hum}
\title{Return partial pressure of water vapor as a function of relative humidity and temperature.}
\usage{
get_vap_pres_from_rel_hum(t_dry_bulb, rel_hum)
}
\arguments{
\item{t_dry_bulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{rel_hum}{Relative humidity in range [0, 1]}
}
\value{
Partial pressure of water vapor in moist air in Psi [IP] or Pa [SI]
}
\description{
Return partial pressure of water vapor as a function of relative humidity and temperature.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 12, 22
}

