% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dew_point_or_relative_humidity_and_vapor_pressure_conversions.R
\name{get_vap_pres_from_t_dew_point}
\alias{get_vap_pres_from_t_dew_point}
\title{Return vapor pressure given dew point temperature.}
\usage{
get_vap_pres_from_t_dew_point(t_dew_point)
}
\arguments{
\item{t_dew_point}{Dew-point temperature in °F [IP] or °C [SI]}
}
\value{
Partial pressure of water vapor in moist air in Psi [IP] or Pa [SI]
}
\description{
Return vapor pressure given dew point temperature.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 36
}

