% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_air.R
\name{GetDryAirEnthalpy}
\alias{GetDryAirEnthalpy}
\title{Return dry-air enthalpy given dry-bulb temperature.}
\usage{
GetDryAirEnthalpy(TDryBulb)
}
\arguments{
\item{TDryBulb}{Dry-bulb temperature in °F [IP] or °C [SI]}
}
\value{
Dry air enthalpy in Btu lb-1 [IP] or J kg-1 [SI]
}
\description{
Return dry-air enthalpy given dry-bulb temperature.
}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 28
}
