% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{GetHumRatioFromSpecificHum}
\alias{GetHumRatioFromSpecificHum}
\title{Return the humidity ratio (aka mixing ratio) from specific humidity.}
\usage{
GetHumRatioFromSpecificHum(SpecificHum)
}
\arguments{
\item{SpecificHum}{Specific humidity in lb_H2O lb_Air-1 [IP] or kg_H2O kg_Air-1 [SI]}
}
\value{
Humidity ratio in lb_H2O lb_Dry_Air-1 [IP] or kg_H2O kg_Dry_Air-1 [SI]
}
\description{
Return the humidity ratio (aka mixing ratio) from specific humidity.
}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 9b (solved for humidity ratio)
}
