% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wet_bulb.R
\name{GetHumRatioFromTDewPoint}
\alias{GetHumRatioFromTDewPoint}
\title{Return humidity ratio given dew-point temperature and pressure.}
\usage{
GetHumRatioFromTDewPoint(TDewPoint, Pressure)
}
\arguments{
\item{TDewPoint}{Dew-point temperature in °F [IP] or °C [SI]}

\item{Pressure}{Atmospheric Pressure in Psi [IP] or Pa [SI]}
}
\value{
Humidity ratio in lb_H2O lb_Air-1 [IP] or kg_H2O kg_Air-1 [SI]
}
\description{
Return humidity ratio given dew-point temperature and pressure.
}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 13
}
