% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wet_bulb.R
\name{GetHumRatioFromTWetBulb}
\alias{GetHumRatioFromTWetBulb}
\title{Return humidity ratio given dry-bulb temperature, wet-bulb temperature, and pressure.}
\usage{
GetHumRatioFromTWetBulb(TDryBulb, TWetBulb, Pressure)
}
\arguments{
\item{TDryBulb}{Dry-bulb temperature in 째F [IP] or 째C [SI]}

\item{TWetBulb}{Wet-bulb temperature in 째F [IP] or 째C [SI]}

\item{Pressure}{Atmospheric Pressure in Psi [IP] or Pa [SI]}
}
\value{
Humidity ratio in lb_H2O lb_Air-1 [IP] or kg_H2O kg_Air-1 [SI]
}
\description{
Return humidity ratio given dry-bulb temperature, wet-bulb temperature, and pressure.
}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 33 and 35
}
