% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapor_press.R
\name{GetHumRatioFromVapPres}
\alias{GetHumRatioFromVapPres}
\title{Return humidity ratio given water vapor Pressure and atmospheric pressure.}
\usage{
GetHumRatioFromVapPres(VapPres, Pressure)
}
\arguments{
\item{VapPres}{Partial Pressure of water vapor in moist air in Psi [IP] or Pa [SI]}

\item{Pressure}{Atmospheric Pressure in Psi [IP] or Pa [SI]}
}
\value{
Humidity ratio in lb_H2O lb_Air-1 [IP] or kg_H2O kg_Air-1 [SI]
}
\description{
Return humidity ratio given water vapor Pressure and atmospheric pressure.
}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 20
}
