% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moist_air.R
\name{GetMoistAirDensity}
\alias{GetMoistAirDensity}
\title{Return moist air density given humidity ratio, dry bulb temperature, and pressure.}
\usage{
GetMoistAirDensity(TDryBulb, HumRatio, Pressure)
}
\arguments{
\item{TDryBulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{HumRatio}{Humidity ratio in lb_H2O lb_Air-1 [IP] or kg_H2O kg_Air-1 [SI]}

\item{Pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
MoistAirDensity: Moist air density in lb ft-3 [IP] or kg m-3 [SI]
}
\description{
Return moist air density given humidity ratio, dry bulb temperature, and pressure.
}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 11
}
