% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wet_bulb.R
\name{GetRelHumFromHumRatio}
\alias{GetRelHumFromHumRatio}
\title{Return relative humidity given dry-bulb temperature, humidity ratio, and pressure.}
\usage{
GetRelHumFromHumRatio(TDryBulb, HumRatio, Pressure)
}
\arguments{
\item{TDryBulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{HumRatio}{Humidity ratio in lb_H2O lb_Air-1 [IP] or kg_H2O kg_Air-1 [SI]}

\item{Pressure}{Atmospheric Pressure in Psi [IP] or Pa [SI]}
}
\value{
Relative humidity in range [0, 1]
}
\description{
Return relative humidity given dry-bulb temperature, humidity ratio, and pressure.
}
\examples{
SetUnitSystem("IP")
GetRelHumFromHumRatio(80:100, 0.01, 14.175)

SetUnitSystem("SI")
GetRelHumFromHumRatio(20:30, 0.01, 95461.0)

}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1
}
