% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_hum.R
\name{GetRelHumFromTDewPoint}
\alias{GetRelHumFromTDewPoint}
\title{Return relative humidity given dry-bulb temperature and dew-point temperature.}
\usage{
GetRelHumFromTDewPoint(TDryBulb, TDewPoint)
}
\arguments{
\item{TDryBulb}{Dry-bulb temperature in 째F [IP] or 째C [SI]}

\item{TDewPoint}{Dew-point temperature in 째F [IP] or 째C [SI]}
}
\value{
Relative humidity in range [0, 1]
}
\description{
Return relative humidity given dry-bulb temperature and dew-point temperature.
}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 22
}
