% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dew_point.R
\name{GetRelHumFromVapPres}
\alias{GetRelHumFromVapPres}
\title{Return relative humidity given dry-bulb temperature and vapor pressure.}
\usage{
GetRelHumFromVapPres(TDryBulb, VapPres)
}
\arguments{
\item{TDryBulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{VapPres}{Partial pressure of water vapor in moist air in Psi [IP] or Pa [SI]}
}
\value{
Relative humidity in range [0, 1]
}
\description{
Return relative humidity given dry-bulb temperature and vapor pressure.
}
\examples{
SetUnitSystem("IP")
GetRelHumFromVapPres(70:80, 0.0149)

SetUnitSystem("SI")
GetRelHumFromVapPres(20:30, 12581)

}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 12, 22
}
