% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_air.R
\name{GetSatAirEnthalpy}
\alias{GetSatAirEnthalpy}
\title{Return saturated air enthalpy given dry-bulb temperature and pressure.}
\usage{
GetSatAirEnthalpy(TDryBulb, Pressure)
}
\arguments{
\item{TDryBulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{Pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Saturated air enthalpy in Btu lb-1 [IP] or J kg-1 [SI]
}
\description{
Return saturated air enthalpy given dry-bulb temperature and pressure.
}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1
}
