% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_air.R
\name{GetSatHumRatio}
\alias{GetSatHumRatio}
\title{Return humidity ratio of saturated air given dry-bulb temperature and pressure.}
\usage{
GetSatHumRatio(TDryBulb, Pressure)
}
\arguments{
\item{TDryBulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{Pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Humidity ratio of saturated air in lb_H2O lb_Air-1 [IP] or kg_H2O kg_Air-1 [SI]
}
\description{
Return humidity ratio of saturated air given dry-bulb temperature and pressure.
}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 36, solved for W
}
