% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmosphere.R
\name{GetSeaLevelPressure}
\alias{GetSeaLevelPressure}
\title{Return sea level pressure given dry-bulb temperature, altitude above sea level and pressure.}
\usage{
GetSeaLevelPressure(StationPressure, Altitude, TDryBulb)
}
\arguments{
\item{StationPressure}{Observed station pressure in Psi [IP] or Pa [SI]}

\item{Altitude}{Altitude in ft [IP] or m [SI]}

\item{TDryBulb}{Dry-bulb temperature in °F [IP] or °C [SI]}
}
\value{
Sea level barometric pressure in Psi [IP] or Pa [SI]
}
\description{
Return sea level pressure given dry-bulb temperature, altitude above sea level and pressure.
}
\note{
The standard procedure for the US is to use for TDryBulb the average
of the current station temperature and the station temperature from 12 hours ago.
}
\references{
Hess SL, Introduction to theoretical meteorology, Holt Rinehart and Winston, NY 1959,
ch. 6.5; Stull RB, Meteorology for scientists and engineers, 2nd edition,
Brooks/Cole 2000, ch. 1.
}
