% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_hum.R
\name{GetTDewPointFromRelHum}
\alias{GetTDewPointFromRelHum}
\title{Return dew-point temperature given dry-bulb temperature and relative humidity.}
\usage{
GetTDewPointFromRelHum(TDryBulb, RelHum)
}
\arguments{
\item{TDryBulb}{Dry-bulb temperature in 째F [IP] or 째C [SI]}

\item{RelHum}{Relative humidity in range [0, 1]}
}
\value{
Dew-point temperature in 째F [IP] or 째C [SI]
}
\description{
Return dew-point temperature given dry-bulb temperature and relative humidity.
}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1
}
