% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_hum.R
\name{GetTDewPointFromTWetBulb}
\alias{GetTDewPointFromTWetBulb}
\title{Return dew-point temperature given dry-bulb temperature, wet-bulb temperature, and pressure.}
\usage{
GetTDewPointFromTWetBulb(TDryBulb, TWetBulb, Pressure)
}
\arguments{
\item{TDryBulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{TWetBulb}{Wet-bulb temperature in °F [IP] or °C [SI]}

\item{Pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Dew-point temperature in °F [IP] or °C [SI]
}
\description{
Return dew-point temperature given dry-bulb temperature, wet-bulb temperature, and pressure.
}
\examples{
SetUnitSystem("IP")
GetTDewPointFromTWetBulb(80:100, 65.0, 14.696)

SetUnitSystem("SI")
GetTDewPointFromTWetBulb(25:40, 20, 101325.0)

}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1
}
