% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_air.R
\name{GetTDryBulbFromEnthalpyAndHumRatio}
\alias{GetTDryBulbFromEnthalpyAndHumRatio}
\title{Return dry bulb temperature from enthalpy and humidity ratio.}
\usage{
GetTDryBulbFromEnthalpyAndHumRatio(MoistAirEnthalpy, HumRatio)
}
\arguments{
\item{MoistAirEnthalpy}{Moist air enthalpy in Btu lb-1 [IP] or J kg-1}

\item{HumRatio}{Humidity ratio in lb_H2O lb_Air-1 [IP] or kg_H2O kg_Air-1 [SI]}
}
\value{
Dry-bulb temperature in °F [IP] or °C [SI]
}
\description{
Return dry bulb temperature from enthalpy and humidity ratio.
}
\note{
Based on the `GetMoistAirEnthalpy` function, rearranged for temperature.
}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 30
}
