% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{GetTKelvinFromTCelsius}
\alias{GetTKelvinFromTCelsius}
\title{Utility function to convert temperature to Kelvin (K)
given temperature in degree Celsius (°C).}
\usage{
GetTKelvinFromTCelsius(TCelsius)
}
\arguments{
\item{TCelsius}{Temperature in degree Celsius (°C)}
}
\value{
Temperature in Kelvin (K)
}
\description{
Utility function to convert temperature to Kelvin (K)
given temperature in degree Celsius (°C).
}
\section{Notes}{

Exact conversion.
}

\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1 section 3
}
