% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{GetTRankineFromTFahrenheit}
\alias{GetTRankineFromTFahrenheit}
\title{Utility function to convert temperature to degree Rankine (°R)
given temperature in degree Fahrenheit (°F).}
\usage{
GetTRankineFromTFahrenheit(TFahrenheit)
}
\arguments{
\item{TFahrenheit}{Temperature in degree Fahrenheit (°F)}
}
\value{
Temperature in degree Rankine (°R)
}
\description{
Utility function to convert temperature to degree Rankine (°R)
given temperature in degree Fahrenheit (°F).
}
\section{Notes}{

Exact conversion.
}

\examples{
GetTRankineFromTFahrenheit(1:100)

}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1 section 3
}
