% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_hum.R
\name{GetTWetBulbFromTDewPoint}
\alias{GetTWetBulbFromTDewPoint}
\title{Return wet-bulb temperature given dry-bulb temperature, dew-point temperature, and pressure.}
\usage{
GetTWetBulbFromTDewPoint(TDryBulb, TDewPoint, Pressure)
}
\arguments{
\item{TDryBulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{TDewPoint}{Dew-point temperature in °F [IP] or °C [SI]}

\item{Pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Wet-bulb temperature in °F [IP] or °C [SI]
}
\description{
Return wet-bulb temperature given dry-bulb temperature, dew-point temperature, and pressure.
}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1
}
