% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{GetUnitSystem}
\alias{GetUnitSystem}
\title{Return system of units in use.}
\usage{
GetUnitSystem()
}
\value{
string indicating system of units in use ("SI" or "IP")
}
\description{
Return system of units in use.
}
