% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moist_air.R
\name{GetVaporPressureDeficit}
\alias{GetVaporPressureDeficit}
\title{Return Vapor pressure deficit given dry-bulb temperature, humidity ratio, and pressure.}
\usage{
GetVaporPressureDeficit(TDryBulb, HumRatio, Pressure)
}
\arguments{
\item{TDryBulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{HumRatio}{Humidity ratio in lb_H2O lb_Air-1 [IP] or kg_H2O kg_Air-1 [SI]}

\item{Pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Vapor pressure deficit in Psi [IP] or Pa [SI]
}
\description{
Return Vapor pressure deficit given dry-bulb temperature, humidity ratio, and pressure.
}
\references{
Oke (1987) eqn 2.13a
}
