% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MAX_ITER_COUNT}
\alias{MAX_ITER_COUNT}
\title{Maximum number of iterations before exiting while loops.}
\format{An object of class \code{numeric} of length 1.}
\usage{
MAX_ITER_COUNT
}
\description{
Maximum number of iterations before exiting while loops.
}
\keyword{datasets}
