% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MIN_HUM_RATIO}
\alias{MIN_HUM_RATIO}
\title{Minimum acceptable humidity ratio used/returned by any functions.
Any value above 0 and below the MIN_HUM_RATIO will be reset to this value.}
\format{An object of class \code{numeric} of length 1.}
\usage{
MIN_HUM_RATIO
}
\description{
Minimum acceptable humidity ratio used/returned by any functions.
Any value above 0 and below the MIN_HUM_RATIO will be reset to this value.
}
\keyword{datasets}
