% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{PSYCHROLIB_TOLERANCE}
\alias{PSYCHROLIB_TOLERANCE}
\title{Tolerance of temperature calculations}
\format{An object of class \code{numeric} of length 1.}
\usage{
PSYCHROLIB_TOLERANCE
}
\description{
Tolerance of temperature calculations
}
\keyword{datasets}
