% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{PSYCHROLIB_UNITS}
\alias{PSYCHROLIB_UNITS}
\title{The system of units in use}
\format{An object of class \code{character} of length 1.}
\usage{
PSYCHROLIB_UNITS
}
\description{
The system of units in use
}
\keyword{datasets}
