% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{PSYCHROLIB_UNITS_OPTIONS}
\alias{PSYCHROLIB_UNITS_OPTIONS}
\title{The options for PSYCHROLIB_UNITS}
\format{An object of class \code{character} of length 2.}
\usage{
PSYCHROLIB_UNITS_OPTIONS
}
\description{
The options for PSYCHROLIB_UNITS
}
\keyword{datasets}
