% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{R_DA_IP}
\alias{R_DA_IP}
\title{Universal gas constant for dry air (IP version)}
\format{An object of class \code{numeric} of length 1.}
\usage{
R_DA_IP
}
\description{
Units:
ft lb_Force lb_DryAir⁻¹ R⁻¹
}
\details{
Reference:
ASHRAE Handbook - Fundamentals (2017) ch. 1
}
\keyword{datasets}
