% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{R_DA_SI}
\alias{R_DA_SI}
\title{Universal gas constant for dry air (SI version)}
\format{An object of class \code{numeric} of length 1.}
\usage{
R_DA_SI
}
\description{
Units:
J kg_DryAir⁻¹ K⁻¹
}
\details{
Reference:
ASHRAE Handbook - Fundamentals (2017) ch. 1
}
\keyword{datasets}
