% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_unit_system}
\alias{set_unit_system}
\title{Set the system of units to use (SI or IP).}
\usage{
set_unit_system(units)
}
\arguments{
\item{units}{string indicating the system of units chosen ("SI" or "IP")

Notes:
This function *HAS TO BE CALLED* before the library can be used}
}
\description{
Set the system of units to use (SI or IP).
}
