% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{SetUnitSystem}
\alias{SetUnitSystem}
\title{Set the system of units to use (SI or IP).}
\usage{
SetUnitSystem(units)
}
\arguments{
\item{units}{string indicating the system of units chosen ("SI" or "IP")}
}
\description{
Set the system of units to use (SI or IP).
}
\note{
The default unit system is "SI".
}
