% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_air_calculations.R
\name{get_dry_air_enthalpy}
\alias{get_dry_air_enthalpy}
\title{Return dry-air enthalpy given dry-bulb temperature.}
\usage{
get_dry_air_enthalpy(t_dry_bulb)
}
\arguments{
\item{t_dry_bulb}{Dry-bulb temperature in °F [IP] or °C [SI]}
}
\value{
Dry air enthalpy in Btu lb⁻¹ [IP] or J kg⁻¹ [SI]
}
\description{
Return dry-air enthalpy given dry-bulb temperature.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 28
}

