% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moist_air_calculations.R
\name{get_moist_air_density}
\alias{get_moist_air_density}
\title{Return moist air density given humidity ratio, dry bulb temperature, and pressure.}
\usage{
get_moist_air_density(t_dry_bulb, hum_ratio, pressure)
}
\arguments{
\item{t_dry_bulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{hum_ratio}{Humidity ratio in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]}

\item{pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
MoistAirDensity: Moist air density in lb ft⁻³ [IP] or kg m⁻³ [SI]
}
\description{
Return moist air density given humidity ratio, dry bulb temperature, and pressure.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 11
}

