% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/wet_bulb_dew_point_or_rel_hum_to_humidity_ratio_conversions.R
\name{get_rel_hum_from_hum_ratio}
\alias{get_rel_hum_from_hum_ratio}
\title{Return relative humidity given dry-bulb temperature, humidity ratio, and pressure.}
\usage{
get_rel_hum_from_hum_ratio(t_dry_bulb, hum_ratio, pressure)
}
\arguments{
\item{t_dry_bulb}{numeric Dry-bulb temperature in °F [IP] or °C [SI]}

\item{hum_ratio}{numeric Humidity ratio in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]}

\item{pressure}{numeric Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
numeric Relative humidity in range [0, 1]

Reference:
  ASHRAE Handbook - Fundamentals (2017) ch. 1
}
\description{
Return relative humidity given dry-bulb temperature, humidity ratio, and pressure.
}
