% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dew_point_or_relative_humidity_and_vapor_pressure_conversions.R
\name{get_rel_hum_from_vap_pres}
\alias{get_rel_hum_from_vap_pres}
\title{Return relative humidity given dry-bulb temperature and vapor pressure.}
\usage{
get_rel_hum_from_vap_pres(t_dry_bulb, vap_pres)
}
\arguments{
\item{t_dry_bulb}{numeric Dry-bulb temperature in °F [IP] or °C [SI]}

\item{vap_pres}{numeric Partial pressure of water vapor in moist air in Psi [IP] or Pa [SI]}
}
\value{
numeric Relative humidity in range [0, 1]

Reference:
  ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 12, 22
}
\description{
Return relative humidity given dry-bulb temperature and vapor pressure.
}
