% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturated_air_calculations.R
\name{get_sat_air_enthalpy}
\alias{get_sat_air_enthalpy}
\title{Return saturated air enthalpy given dry-bulb temperature and pressure.}
\usage{
get_sat_air_enthalpy(t_dry_bulb, pressure)
}
\arguments{
\item{t_dry_bulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Saturated air enthalpy in Btu lb⁻¹ [IP] or J kg⁻¹ [SI]
}
\description{
Return saturated air enthalpy given dry-bulb temperature and pressure.
}
\section{Reference}{

 ASHRAE Handbook - Fundamentals (2017) ch. 1
}

