% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_atmosphere_calculations.R
\name{get_station_pressure}
\alias{get_station_pressure}
\title{Return station pressure from sea level pressure.}
\usage{
get_station_pressure(sea_level_pressure, altitude, t_dry_bulb)
}
\arguments{
\item{sea_level_pressure}{numeric Sea level barometric pressure in Psi [IP] or Pa [SI]}

\item{altitude}{numeric Altitude in ft [IP] or m [SI]}

\item{t_dry_bulb}{numeric Dry-bulb temperature in °F [IP] or °C [SI]}
}
\value{
numeric Station pressure in Psi [IP] or Pa [SI]

Reference:
  See 'GetSeaLevelPressure'

Notes:
  This function is just the inverse of 'GetSeaLevelPressure'.
}
\description{
Return station pressure from sea level pressure.
}
