% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{get_t_kelvin_from_t_celsius}
\alias{get_t_kelvin_from_t_celsius}
\title{Utility function to convert temperature to Kelvin (K)
given temperature in degree Celsius (°C).}
\usage{
get_t_kelvin_from_t_celsius(t_celsius)
}
\arguments{
\item{t_celsius}{numeric Temperature in degree Celsius (°C)}
}
\value{
numeric Temperature in Kelvin (K)

Notes:
  Exact conversion.
}
\description{
Utility function to convert temperature to Kelvin (K)
given temperature in degree Celsius (°C).
}
