% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dew_point_wet_bulb_relative_humidity_conversions.R
\name{get_t_wet_bulb_from_t_dew_point}
\alias{get_t_wet_bulb_from_t_dew_point}
\title{Return wet-bulb temperature given dry-bulb temperature, dew-point temperature, and pressure.}
\usage{
get_t_wet_bulb_from_t_dew_point(t_dry_bulb, t_dew_point, pressure)
}
\arguments{
\item{t_dry_bulb}{Dry-bulb temperature in °F [IP] or °C [SI]}

\item{t_dew_point}{Dew-point temperature in °F [IP] or °C [SI]}

\item{pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Wet-bulb temperature in °F [IP] or °C [SI]
}
\description{
Return wet-bulb temperature given dry-bulb temperature, dew-point temperature, and pressure.
}
\section{Reference}{

  ASHRAE Handbook - Fundamentals (2017) ch. 1
}

