% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moist_air_calculations.R
\name{get_vapor_pressure_deficit}
\alias{get_vapor_pressure_deficit}
\title{Return Vapor pressure deficit given dry-bulb temperature, humidity ratio, and pressure.}
\usage{
get_vapor_pressure_deficit(t_dry_bulb, hum_ratio, pressure)
}
\arguments{
\item{t_dry_bulb}{numeric Dry-bulb temperature in °F [IP] or °C [SI]}

\item{hum_ratio}{numeric Humidity ratio in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]}

\item{pressure}{numeric Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
numeric Vapor pressure deficit in Psi [IP] or Pa [SI]

Reference:
  Oke (1987) eqn 2.13a
}
\description{
Return Vapor pressure deficit given dry-bulb temperature, humidity ratio, and pressure.
}
