% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{GetTCelsiusFromTKelvin}
\alias{GetTCelsiusFromTKelvin}
\title{Utility function to convert temperature to degree Celsius (°C)
given temperature in Kelvin (K).}
\usage{
GetTCelsiusFromTKelvin(TKelvin)
}
\arguments{
\item{TKelvin}{Temperature in degree Kelvin (K)}
}
\value{
Temperature in Celsius (°C)
}
\description{
Utility function to convert temperature to degree Celsius (°C)
given temperature in Kelvin (K).
}
\section{Notes}{

Exact conversion.
}

\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1 section 3
}
