% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wet_bulb.R
\name{GetTDewPointFromHumRatio}
\alias{GetTDewPointFromHumRatio}
\title{Return dew-point temperature given dry-bulb temperature, humidity ratio, and pressure.}
\usage{
GetTDewPointFromHumRatio(TDryBulb, HumRatio, Pressure)
}
\arguments{
\item{TDryBulb}{Dry-bulb temperature in 째F [IP] or 째C [SI]}

\item{HumRatio}{Humidity ratio in lb_H2O lb_Air-1 [IP] or kg_H2O kg_Air-1 [SI]}

\item{Pressure}{Atmospheric Pressure in Psi [IP] or Pa [SI]}
}
\value{
Dew-point temperature in 째F [IP] or 째C [SI]
}
\description{
Return dew-point temperature given dry-bulb temperature, humidity ratio, and pressure.
}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1
}
