% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrolib.R
\name{GetTDewPointFromHumRatio}
\alias{GetTDewPointFromHumRatio}
\title{Return dew-point temperature given dry-bulb temperature, humidity ratio, and pressure.}
\usage{
GetTDewPointFromHumRatio(TDryBulb, HumRatio, Pressure)
}
\arguments{
\item{TDryBulb}{A numeric vector of dry-bulb temperature in degreeF [IP] or degreeC [SI]}

\item{HumRatio}{A numeric vector of humidity ratio in lb_H2O lb_Air-1 [IP] or kg_H2O kg_Air-1 [SI]}

\item{Pressure}{A numeric vector of atmospheric Pressure in Psi [IP] or Pa [SI]}
}
\value{
A numeric vector of dew-point temperature in degreeF [IP] or degreeC [SI]
}
\description{
Return dew-point temperature given dry-bulb temperature, humidity ratio, and pressure.
}
\examples{
SetUnitSystem("IP")
GetTDewPointFromHumRatio(80:100, 0.01, 14.175)

SetUnitSystem("SI")
GetTDewPointFromHumRatio(20:30, 0.01, 95461.0)

}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1
}
