% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrolib.R
\name{GetTFahrenheitFromTRankine}
\alias{GetTFahrenheitFromTRankine}
\title{Utility function to convert temperature to degree Fahrenheit (degreeF)
given temperature in degree Rankine (degreeR).}
\usage{
GetTFahrenheitFromTRankine(TRankine)
}
\arguments{
\item{TRankine}{A numeric vector of temperature in degree Rankine (degreeR)}
}
\value{
A numeric vector of temperature in degree Fahrenheit (degreeF)
}
\description{
Utility function to convert temperature to degree Fahrenheit (degreeF)
given temperature in degree Rankine (degreeR).
}
\section{Notes}{

Exact conversion.
}

\examples{
GetTFahrenheitFromTRankine(500:600)

}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1 section 3
}
