% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrolib.R
\name{GetTKelvinFromTCelsius}
\alias{GetTKelvinFromTCelsius}
\title{Utility function to convert temperature to Kelvin (K)
given temperature in degree Celsius (degreeC).}
\usage{
GetTKelvinFromTCelsius(TCelsius)
}
\arguments{
\item{TCelsius}{A numeric vector of temperature in degree Celsius (degreeC)}
}
\value{
A numeric vector of temperature in Kelvin (K)
}
\description{
Utility function to convert temperature to Kelvin (K)
given temperature in degree Celsius (degreeC).
}
\section{Notes}{

Exact conversion.
}

\examples{
GetTKelvinFromTCelsius(20:30)

}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1 section 3
}
