% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrolib.R
\name{GetTRankineFromTFahrenheit}
\alias{GetTRankineFromTFahrenheit}
\title{Utility function to convert temperature to degree Rankine (degreeR)
given temperature in degree Fahrenheit (degreeF).}
\usage{
GetTRankineFromTFahrenheit(TFahrenheit)
}
\arguments{
\item{TFahrenheit}{A numeric vector of temperature in degree Fahrenheit (degreeF)}
}
\value{
A numeric vector of temperature in degree Rankine (degreeR)
}
\description{
Utility function to convert temperature to degree Rankine (degreeR)
given temperature in degree Fahrenheit (degreeF).
}
\section{Notes}{

Exact conversion.
}

\examples{
GetTRankineFromTFahrenheit(1:100)

}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1 section 3
}
