% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_hum.R
\name{GetTWetBulbFromRelHum}
\alias{GetTWetBulbFromRelHum}
\title{Return wet-bulb temperature given dry-bulb temperature, relative humidity, and pressure.}
\usage{
GetTWetBulbFromRelHum(TDryBulb, RelHum, Pressure)
}
\arguments{
\item{TDryBulb}{Dry-bulb temperature in 째F [IP] or 째C [SI]}

\item{RelHum}{Relative humidity in range [0, 1]}

\item{Pressure}{Atmospheric pressure in Psi [IP] or Pa [SI]}
}
\value{
Wet-bulb temperature in 째F [IP] or 째C [SI]
}
\description{
Return wet-bulb temperature given dry-bulb temperature, relative humidity, and pressure.
}
\examples{
SetUnitSystem("IP")
GetTWetBulbFromRelHum(80:100, 0.2, 14.696)

SetUnitSystem("SI")
GetTWetBulbFromRelHum(25:40, 0.2, 101325.0)

}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1
}
