% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapor_press.R
\name{GetVapPresFromHumRatio}
\alias{GetVapPresFromHumRatio}
\title{Return vapor Pressure given humidity ratio and pressure.}
\usage{
GetVapPresFromHumRatio(HumRatio, Pressure)
}
\arguments{
\item{HumRatio}{Humidity ratio in lb_H2O lb_Air-1 [IP] or kg_H2O kg_Air-1 [SI]}

\item{Pressure}{Atmospheric Pressure in Psi [IP] or Pa [SI]}
}
\value{
Partial Pressure of water vapor in moist air in Psi [IP] or Pa [SI]
}
\description{
Return vapor Pressure given humidity ratio and pressure.
}
\references{
ASHRAE Handbook - Fundamentals (2017) ch. 1 eqn 20 solved for pw
}
