% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychro-package.R
\docType{package}
\name{psychrolib-package}
\alias{psychrolib}
\alias{psychrolib-package}
\title{psychrolib: Psychrometric Properties of Moist and Dry Air}
\description{
Psychrometrics are the study of physical and thermodynamic properties of
moist air. These properties include, for example, the air's dew point
temperature, its wet bulb temperature, relative humidity, humidity ratio,
enthalpy.
}
\details{
The estimation of these properties is critical in several engineering and
scientific applications such as heating, ventilation, and air conditioning
(HVAC) and meteorology.

psychroLib is a port of the
\href{https://github.com/psychrometrics/psychrolib}{psychrolib} library for
R. It provides functions to enable calculating psychrometric properties of
moist and dry air, working in both metric (SI) and imperial (IP) systems of
units. The functions are based of formulae from the 2017 ASHRAE Handbook —
Fundamentals, Chapter 1, SI and IP editions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/psychrometrics/psychrolib}
  \item Report bugs at \url{https://github.com/psychrometrics/psychrolib/issues}
}

}
\author{
\itemize{
  \item Hongyuan Jia and Jason Banfelder for R implementation.
  \item D. Thevenard and D. Meyer for the core library implementations.
}
}
