% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{SetUnitSystem}
\alias{SetUnitSystem}
\title{Set the system of units to use (SI or IP).}
\usage{
SetUnitSystem(units)
}
\arguments{
\item{units}{A string indicating the system of units chosen. Should be either
\code{"SI"} or \code{"IP"}}
}
\description{
Set the system of units to use (SI or IP).
}
